Working toward proper space integration.
- Starting by moving all logic from game_hack to more proper locations. Then I'll worry about
    fixing the logic.
- Moved drawing from game_hack to graphics_system.
- Working towards transforming game_hack into a scene.
- Added operator<< for vecn.
- Added owner storing and set/getters for component, entity, space, and game_hack_scene.

+++ src/engine/window.cpp
@@ -35,11 +35,11 @@ void window::update()
 }
 
 // TODO: Cache size
-ivec2 window::get_size() const
+uvec2 window::get_size() const
 {
   int w, h;
   SDL_GetWindowSize(p_raw_window, &w, &h);
-  return ivec2(w, h);
+  return uvec2(static_cast<unsigned>(w), static_cast<unsigned>(h));
 }
 
 SDL_Window const* window::get_raw() const
